Can now create Booking-object which uses User and Event-object ID's as attributes. Possible to fetch information about the Booking User-Object and the Booking Event-Object.
Booking-Object now takes in ID of User and Event as FK to be able to be able to create efficient queries and and still be able to filter. But also includes User and Event Object as attributes to enable Lazy/Eager Loading which is powerful in scenarios where the full objects are needed.